pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elt-observability"
version = "0.1.0"
description = "Production-ready ELT Observability Platform for dbt + Airflow stacks"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.25",
    "structlog>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["collector", "detectors", "storage", "dashboards", "alerting"]
exclude = ["tests", "output", "dbt_integration"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--tb=short -v"
